+++ frontend/src/components/Login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import style from '../Login.module.css'
+import style from '../css/Login.module.css'
 import Axios from 'axios';
 class Login extends React.Component{
     constructor(props) {
