+++ frontend/src/css/Registration.module.css
